[Refactor] Break part of `VictoryPhase` into `PostKnockoutPhase`
