+++ src/phases/post-knockout-phase.ts
@@ -1,0 +1,55 @@
+import { handleMysteryEncounterVictory } from "#app/data/mystery-encounters/utils/encounter-phase-utils";
+import { globalScene } from "#app/global-scene";
+import { PokemonPhase } from "#app/phases/abstract-pokemon-phase";
+import { VictoryPhase } from "#app/phases/victory-phase";
+import { BattleType } from "#enums/battle-type";
+import type { BattlerIndex } from "#enums/battler-index";
+import { PhaseId } from "#enums/phase-id";
+
+/**
+ * Handles the actions after the player KOs a pokemon:
+ * - Increases {@linkcode globalScene.gameData.gameStats.pokemonDefeated | pokemon defeated} count in game stats
+ * - Applies EXP gain
+ * - If this is a Mystery Encounter: hand off to a ME function and end the phase
+ * - If there are no more unfainted pokemon on the enemy team, unshift a {@linkcode VictoryPhase}
+ */
+export class PostKnockoutPhase extends PokemonPhase {
+  public override readonly id: PhaseId = PhaseId.POST_KNOCKOUT;
+  /**
+   * If `true`, indicates that the phase is intended for EXP purposes only, and not to continue a battle to next phase.
+   * Only used by Mystery Encounters.
+   */
+  public readonly isExpOnly: boolean;
+
+  constructor(battlerIndex: BattlerIndex | number, isExpOnly: boolean = false) {
+    super(battlerIndex);
+
+    this.isExpOnly = isExpOnly;
+  }
+
+  public override start(): void {
+    const { currentBattle, gameData, phaseManager } = globalScene;
+    const { battleType, mysteryEncounter } = currentBattle;
+
+    const isMysteryEncounter = currentBattle.isBattleMysteryEncounter();
+
+    // update Pokemon defeated count except for MEs that disable it
+    if (!isMysteryEncounter || !mysteryEncounter?.preventGameStatsUpdates) {
+      gameData.gameStats.pokemonDefeated++;
+    }
+
+    const expValue = this.getPokemon().getExpValue();
+    globalScene.applyPartyExp(expValue, true);
+
+    if (isMysteryEncounter) {
+      handleMysteryEncounterVictory(false, this.isExpOnly);
+      return this.end();
+    }
+
+    if (!globalScene.getEnemyParty().some((p) => (battleType === BattleType.WILD ? p.isOnField() : !p?.isFainted()))) {
+      phaseManager.unshiftPhase(new VictoryPhase(this.battlerIndex));
+    }
+
+    this.end();
+  }
+}
