+++ src/phases/victory-phase.ts
@@ -4,7 +4,6 @@ import type { GameOverPhase } from "#app/phases/game-over-phase";
 /* eslint-enable @typescript-eslint/no-unused-vars */
 // -- end tsdoc imports--
 
-import { handleMysteryEncounterVictory } from "#app/data/mystery-encounters/utils/encounter-phase-utils";
 import { EVIL_BOSS_2_WAVE } from "#app/data/special-waves";
 import { globalScene } from "#app/global-scene";
 import { type CustomModifierSettings } from "#app/modifier/modifier-type";
@@ -22,12 +21,7 @@ import { type BattlerIndex } from "#enums/battler-index";
 import { PhaseId } from "#enums/phase-id";
 
 /**
- * Handles the actions after the player KOs a pokemon:
- * - Increases {@linkcode globalScene.gameData.gameStats.pokemonDefeated | pokemon defeated} count in game stats
- * - Applies EXP gain
- * - Ends the phase early:
- *   - If this is a Mystery Encounter, in which case execution is handed off to a ME function; or
- *   - If there are still unfainted pokemon on the enemy team
+ * Handles various effects when the player clears a wave:
  * - Clears "delayed attack" arena tags (Future Sight / Doom Desire)
  * - Pushes a {@linkcode BattleEndPhase}
  * - Pushes a {@linkcode TrainerVictoryPhase} if this was a trainer battle
@@ -38,43 +32,17 @@ import { PhaseId } from "#enums/phase-id";
  */
 export class VictoryPhase extends PokemonPhase {
   override readonly id = PhaseId.VICTORY;
-  /**
-   * If `true`, indicates that the phase is intended for EXP purposes only, and not to continue a battle to next phase.
-   * Only used by Mystery Encounters.
-   */
-  public readonly isExpOnly: boolean;
 
-  constructor(battlerIndex: BattlerIndex | number, isExpOnly: boolean = false) {
+  constructor(battlerIndex: BattlerIndex | number) {
     super(battlerIndex);
-
-    this.isExpOnly = isExpOnly;
   }
 
   public override start(): void {
     super.start();
 
-    const { currentBattle, gameData, gameMode, phaseManager } = globalScene;
-    const { battleType, mysteryEncounter, waveIndex } = currentBattle;
+    const { currentBattle, gameMode, phaseManager } = globalScene;
+    const { battleType, waveIndex } = currentBattle;
     const { isClassic, isDaily, isEndless } = gameMode;
-
-    const isMysteryEncounter = currentBattle.isBattleMysteryEncounter();
-
-    // update Pokemon defeated count except for MEs that disable it
-    if (!isMysteryEncounter || !mysteryEncounter?.preventGameStatsUpdates) {
-      gameData.gameStats.pokemonDefeated++;
-    }
-
-    const expValue = this.getPokemon().getExpValue();
-    globalScene.applyPartyExp(expValue, true);
-
-    if (isMysteryEncounter) {
-      handleMysteryEncounterVictory(false, this.isExpOnly);
-      return super.end();
-    }
-
-    if (globalScene.getEnemyParty().some((p) => (battleType === BattleType.WILD ? p.isOnField() : !p?.isFainted()))) {
-      return super.end();
-    }
 
     // clear all queued delayed attacks (e.g. from Future Sight)
     globalScene.arena.removeTag(ArenaTagType.DELAYED_ATTACK);
