+++ test/mystery-encounter/encounter-test-utils.ts
@@ -2,17 +2,17 @@ import * as EncounterPhaseUtils from "#app/data/mystery-encounters/utils/encount
 import { CommandPhase } from "#app/phases/command-phase";
 import { MessagePhase } from "#app/phases/message-phase";
 import { MysteryEncounterBattlePhase } from "#app/phases/mystery-encounter-phases/battle-phase";
-import { MysteryEncounterRewardsPhase } from "#app/phases/mystery-encounter-phases/rewards-phase";
-import { MysteryEncounterOptionSelectedPhase } from "#app/phases/mystery-encounter-phases/option-selected-phase";
 import { MysteryEncounterPhase } from "#app/phases/mystery-encounter-phases/mystery-encounter-phase";
-import { VictoryPhase } from "#app/phases/victory-phase";
+import { MysteryEncounterOptionSelectedPhase } from "#app/phases/mystery-encounter-phases/option-selected-phase";
+import { MysteryEncounterRewardsPhase } from "#app/phases/mystery-encounter-phases/rewards-phase";
+import { PostKnockoutPhase } from "#app/phases/post-knockout-phase";
 import type { MessageUiHandler } from "#app/ui/handlers/message-ui-handler";
 import type { MysteryEncounterUiHandler } from "#app/ui/handlers/mystery-encounter-ui-handler";
-import type { PartyUiHandler } from "#app/ui/handlers/party-ui-handler";
 import type { OptionSelectUiHandler } from "#app/ui/handlers/option-select-ui-handler";
-import { UiMode } from "#enums/ui-mode";
+import type { PartyUiHandler } from "#app/ui/handlers/party-ui-handler";
 import { isNullOrUndefined } from "#app/utils";
 import { Button } from "#enums/buttons";
+import { UiMode } from "#enums/ui-mode";
 import type { GameManager } from "#test/test-utils/gameManager";
 import { expect, vi } from "vitest";
 
@@ -68,7 +68,7 @@ export async function runMysteryEncounterToEnd(
     game.onNextPrompt("CommandPhase", UiMode.COMMAND, () => {
       game.scene.phaseManager.clearPhaseQueue();
       game.scene.phaseManager.clearPhaseQueueSplice();
-      game.scene.phaseManager.unshiftPhase(new VictoryPhase(0));
+      game.scene.phaseManager.unshiftPhase(new PostKnockoutPhase(0));
       game.endPhase();
     });
 
@@ -199,7 +199,7 @@ export async function skipBattleRunMysteryEncounterRewardsPhase(game: GameManage
     p.faint();
     game.scene.field.remove(p);
   });
-  game.scene.phaseManager.pushPhase(new VictoryPhase(0));
+  game.scene.phaseManager.pushPhase(new PostKnockoutPhase(0));
   game.phaseInterceptor.superEndPhase();
   game.setMode(UiMode.MESSAGE);
   await game.phaseInterceptor.to(MysteryEncounterRewardsPhase, runRewardsPhase);
